bug: pagination not working
The previous changes used a page number of -1 if there were remaining pages to be rendered on the next AJAX call from the repository API.
This is not the correct way to handle pages with the repository API and was causing only the first page to be rendered in some cases.
Fixed pagination by passing full page count to the listing.